Contour extraction.py
Given an image we can segregate the objects witth respect to there edges / boundries ie. contours. 
First we convert image from RGB to BGR as OpenCV reads image in BGR form. Then we convert
BGR image to gray so, it can be further converted to binary image. 
The main reason to convert an image to binary form is to seperate the object and background pixels into white and black respectively.
Finally we use a module called "Find Contours" on the binary image which returns the plotted contour on original colored image.